idoc-backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.4.0
django-allauth==65.3.0

# Database
psycopg2-binary==2.9.10
django-environ==0.11.2

# Real-time
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Payments
stripe==11.4.1

# Video (Agora token generation)
agora-token-builder==1.0.0

# File handling
Pillow==11.1.0
django-storages==1.14.4
boto3==1.35.86

# Utilities
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
gunicorn==23.0.0
whitenoise==6.8.2

# Push Notifications
firebase-admin==6.6.0

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3
